humanloop_backend/requirements.txt
# Django Framework
Django==4.2.28

# Database
mysqlclient==2.2.8

# Image Processing
Pillow==12.1.1

# Two-Factor Authentication
pyotp==2.9.0
qrcode==8.2

# Payment Gateways
razorpay==2.0.0
stripe==14.4.0

# HTTP Requests (for AI/external APIs)
requests==2.32.5

# Environment Variables
python-dotenv==1.1.0
